requirements.txt
easysettings
python-twitter
requests
sh
